tests: extract SpanContextMock file, change type of PropagatedTags
